+++ src/main/java/org/folio/marc/migrations/services/operations/ChunkService.java
@@ -12,9 +12,11 @@ import lombok.extern.log4j.Log4j2;
 import org.folio.marc.migrations.config.MigrationProperties;
 import org.folio.marc.migrations.domain.entities.Operation;
 import org.folio.marc.migrations.domain.entities.OperationChunk;
+import org.folio.marc.migrations.domain.entities.types.EntityType;
 import org.folio.marc.migrations.domain.entities.types.OperationStatusType;
 import org.folio.marc.migrations.services.jdbc.AuthorityJdbcService;
 import org.folio.marc.migrations.services.jdbc.ChunkJdbcService;
+import org.folio.marc.migrations.services.jdbc.InstanceJdbcService;
 import org.springframework.stereotype.Service;
 
 @Log4j2
@@ -27,17 +29,22 @@ public class ChunkService {
   private final MigrationProperties props;
   private final AuthorityJdbcService authorityJdbcService;
   private final ChunkJdbcService chunkJdbcService;
+  private final InstanceJdbcService instanceJdbcService;
 
   public void prepareChunks(Operation operation) {
     log.info("prepareChunks:: starting for operation {}", operation.getId());
     var chunks = new LinkedList<OperationChunk>();
 
-    var recordIds = authorityJdbcService.getAuthorityIdsChunk(props.getChunkFetchIdsCount());
+    var recordIds = (operation.getEntityType() == EntityType.AUTHORITY) ?
+        authorityJdbcService.getAuthorityIdsChunk(props.getChunkFetchIdsCount()):
+        instanceJdbcService.getInstanceIdsChunk(props.getChunkFetchIdsCount());
     addChunksForRecordIds(operation, chunks, recordIds);
     var idFrom = Optional.ofNullable(chunks.peekLast()).map(OperationChunk::getEndRecordId).orElse(null);
 
     while (recordIds.size() == props.getChunkFetchIdsCount()) {
-      recordIds = authorityJdbcService.getAuthorityIdsChunk(idFrom, props.getChunkFetchIdsCount());
+      recordIds = (operation.getEntityType() == EntityType.AUTHORITY) ?
+          authorityJdbcService.getAuthorityIdsChunk(idFrom, props.getChunkFetchIdsCount()):
+          instanceJdbcService.getInstanceIdsChunk(idFrom, props.getChunkFetchIdsCount());
       addChunksForRecordIds(operation, chunks, recordIds);
 
       idFrom = Optional.ofNullable(chunks.peekLast()).map(OperationChunk::getEndRecordId).orElse(null);