+++ src/main/java/org/folio/marc/migrations/services/tenant/ModuleTenantService.java
@@ -2,6 +2,7 @@ package org.folio.marc.migrations.services.tenant;
 
 import lombok.extern.log4j.Log4j2;
 import org.folio.marc.migrations.services.jdbc.AuthorityJdbcService;
+import org.folio.marc.migrations.services.jdbc.InstanceJdbcService;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.liquibase.FolioSpringLiquibase;
 import org.folio.spring.service.PrepareSystemUserService;
@@ -18,14 +19,16 @@ public class ModuleTenantService extends TenantService {
 
   private final PrepareSystemUserService systemUserService;
   private final AuthorityJdbcService authorityJdbcService;
+  private final InstanceJdbcService instanceJdbcService;
 
   public ModuleTenantService(JdbcTemplate jdbcTemplate,
                              FolioExecutionContext context,
                              FolioSpringLiquibase folioSpringLiquibase,
-                             PrepareSystemUserService systemUserService, AuthorityJdbcService authorityJdbcService) {
+                             PrepareSystemUserService systemUserService, AuthorityJdbcService authorityJdbcService, InstanceJdbcService instanceJdbcService) {
     super(jdbcTemplate, context, folioSpringLiquibase);
     this.systemUserService = systemUserService;
     this.authorityJdbcService = authorityJdbcService;
+    this.instanceJdbcService = instanceJdbcService;
   }
 
   @Override
@@ -34,6 +37,7 @@ public class ModuleTenantService extends TenantService {
     super.afterTenantUpdate(tenantAttributes);
     systemUserService.setupSystemUser();
     authorityJdbcService.initViews(context.getTenantId());
+    instanceJdbcService.initViews(context.getTenantId());
     log.info("afterTenantUpdate::Completed additional setup [tenant: {}]", context.getTenantId());
   }
 }